+++ backend/setup/settings.py
@@ -40,7 +40,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'corsheaders',
     'rest_framework',
-    'backend.api'
+    'backend.core.api'
 ]
 
 MIDDLEWARE = [
